second commit

+++ js/index.js
@@ -68,9 +68,26 @@ function clear_inputs() {
 }
 
 function clear_item(index) {
-  array_list.splice(index, 1);
-  localStorage.setItem("product", JSON.stringify(array_list));
-  display_Card();
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      array_list.splice(index, 1);
+      localStorage.setItem("product", JSON.stringify(array_list));
+      display_Card();
+      Swal.fire({
+        title: "Deleted!",
+        text: "Your file has been deleted.",
+        icon: "success",
+      });
+    }
+  });
 }
 function recale_Item(index) {
   input_ProductName.value = array_list[index].name;
